% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long-to-preference.R
\name{long_to_preference}
\alias{long_to_preference}
\title{Convert a long-form ranks into a preference matrix.}
\usage{
long_to_preference(d_rank_college, d_rank_student)
}
\arguments{
\item{d_rank_college}{How the colleges ranked the students.  A 1 represents a college's top choice.}

\item{d_rank_student}{How the students ranked the colleges.  A 1 represents a student's top choice.}
}
\description{
A long-format dataset (where each row is a specific rank) for the college and for the students
are converted to a wide-format preference matrix.
}
\author{
Will Beasley
}
